refactor: project restructuring
This commit moves all cobra functions to their respective files,
strictly following their best-prectice guides.
Close #83
